Breaking: rename CollapseElement to CollapseTable
Try to limit scope by renaming CollapseElement to CollapseTable
